Funcionalidad de la aplicación 100%. Fix bug a la hora de buscar un Movimiento de x articulo. Tema de la App modificado. Ahora en el Tiempo te indica el error correctamente. Si no hay movimientos te lo indica. Ahora a la hora de eliminar un articulo deslizando arreglado y de la manera tradicional o deslizando te indica el codigo y la descripción de ese articulo.

+++ app/src/main/java/com/sergigonzalez/gestionarticulos/ui/fragments/FragmentMain.kt
@@ -1,23 +1,25 @@
 package com.sergigonzalez.gestionarticulos.ui.fragments
 
+import android.app.AlertDialog
+import android.content.Context
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import android.view.*
 import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.sergigonzalez.gestionarticulos.R
 import com.sergigonzalez.gestionarticulos.adapters.ArticleAdapter
 import com.sergigonzalez.gestionarticulos.data.Article
 import com.sergigonzalez.gestionarticulos.data.ArticleApp
 import com.sergigonzalez.gestionarticulos.databinding.FragmentMainBinding
 
+
 class FragmentMain : Fragment() {
     private lateinit var database: ArticleApp
     private var listArticles: List<Article> = emptyList()
-    private var _binding:FragmentMainBinding? = null
+    private var _binding: FragmentMainBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -59,13 +61,26 @@ class FragmentMain : Fragment() {
 
                 override fun onSwiped(viewHolder: RecyclerView.ViewHolder, swipeDir: Int) {
                     //3
+                    val builder = AlertDialog.Builder(view.context)
                     val position = viewHolder.adapterPosition
-                    ArticleAdapter.ArticleHolder.deleteArticle(
-                        listArticles[position],
-                        binding.root,
-                        database
-                    )
-                    binding.lista.adapter!!.notifyItemRemoved(position)
+                    builder.setMessage("Estas seguro que deseas eliminar el Articulo?\nCon codigo: ${listArticles[position].idArticle}\nDescripción:  ${listArticles[position].descriptionArticle}")
+
+
+                    builder.setPositiveButton(android.R.string.ok) { _, _ ->
+                        ArticleAdapter.ArticleHolder.deleteArticle(
+                            listArticles[position],
+                            database
+                        )
+                        binding.lista.adapter!!.notifyItemRemoved(position)
+                    }
+
+                    builder.setNegativeButton(android.R.string.cancel) { dialog, _ ->
+                        dialog.dismiss()
+                        binding.lista.adapter!!.notifyItemChanged(position)
+                    }
+
+                    builder.show()
+
                 }
             }
 
@@ -86,6 +101,181 @@ class FragmentMain : Fragment() {
         })
     }
 
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        inflater.inflate(R.menu.menu_main, menu)
+        super.onCreateOptionsMenu(menu, inflater)
+    }
+
+    override fun onAttach(context: Context) {
+        setHasOptionsMenu(true);
+        super.onAttach(context)
+    }
+
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        when (item.itemId) {
+
+            //All filters
+            R.id.AllArticles -> {
+                database.Articles().getAll().observe(this, {
+                    listArticles = it
+                    binding.lista.layoutManager =
+                        LinearLayoutManager(this@FragmentMain.requireContext())
+                    binding.lista.addItemDecoration(
+                        DividerItemDecoration(
+                            this@FragmentMain.requireContext(),
+                            DividerItemDecoration.VERTICAL
+                        )
+                    )
+
+                    val adapter = ArticleAdapter(listArticles)
+
+                    binding.lista.adapter = adapter
+                    if (listArticles.isEmpty()) {
+                        binding.emptyView.visibility = View.VISIBLE
+                        binding.lista.visibility = View.GONE
+                    } else {
+                        binding.emptyView.visibility = View.GONE
+                        binding.lista.visibility = View.VISIBLE
+                    }
+
+                })
+            }
+
+            R.id.ArticlesWithDescription -> {
+                database.Articles().getDescription().observe(this, {
+                    listArticles = it
+                    binding.lista.layoutManager =
+                        LinearLayoutManager(this@FragmentMain.requireContext())
+                    binding.lista.addItemDecoration(
+                        DividerItemDecoration(
+                            this@FragmentMain.requireContext(),
+                            DividerItemDecoration.VERTICAL
+                        )
+                    )
+
+                    val adapter = ArticleAdapter(listArticles)
+
+                    binding.lista.adapter = adapter
+                    if (listArticles.isEmpty()) {
+                        binding.emptyView.visibility = View.VISIBLE
+                        binding.lista.visibility = View.GONE
+                    } else {
+                        binding.emptyView.visibility = View.GONE
+                        binding.lista.visibility = View.VISIBLE
+                    }
+
+                })
+            }
+
+            R.id.NoHaveStock -> {
+                database.Articles().getWithoutStock().observe(this, {
+                    listArticles = it
+                    binding.lista.layoutManager =
+                        LinearLayoutManager(this@FragmentMain.requireContext())
+                    binding.lista.addItemDecoration(
+                        DividerItemDecoration(
+                            this@FragmentMain.requireContext(),
+                            DividerItemDecoration.VERTICAL
+                        )
+                    )
+
+                    val adapter = ArticleAdapter(listArticles)
+
+                    binding.lista.adapter = adapter
+                    if (listArticles.isEmpty()) {
+                        binding.emptyView.visibility = View.VISIBLE
+                        binding.lista.visibility = View.GONE
+                    } else {
+                        binding.emptyView.visibility = View.GONE
+                        binding.lista.visibility = View.VISIBLE
+                    }
+
+                })
+
+            }
+
+            R.id.HaveStock -> {
+                database.Articles().getWithStock().observe(this, {
+                    listArticles = it
+                    binding.lista.layoutManager =
+                        LinearLayoutManager(this@FragmentMain.requireContext())
+                    binding.lista.addItemDecoration(
+                        DividerItemDecoration(
+                            this@FragmentMain.requireContext(),
+                            DividerItemDecoration.VERTICAL
+                        )
+                    )
+
+                    val adapter = ArticleAdapter(listArticles)
+
+                    binding.lista.adapter = adapter
+                    if (listArticles.isEmpty()) {
+                        binding.emptyView.visibility = View.VISIBLE
+                        binding.lista.visibility = View.GONE
+                    } else {
+                        binding.emptyView.visibility = View.GONE
+                        binding.lista.visibility = View.VISIBLE
+                    }
+
+                })
+            }
+
+            R.id.Ascendent -> {
+                database.Articles().getAllAsc().observe(this, {
+                    listArticles = it
+                    binding.lista.layoutManager =
+                        LinearLayoutManager(this@FragmentMain.requireContext())
+                    binding.lista.addItemDecoration(
+                        DividerItemDecoration(
+                            this@FragmentMain.requireContext(),
+                            DividerItemDecoration.VERTICAL
+                        )
+                    )
+
+                    val adapter = ArticleAdapter(listArticles)
+
+                    binding.lista.adapter = adapter
+                    if (listArticles.isEmpty()) {
+                        binding.emptyView.visibility = View.VISIBLE
+                        binding.lista.visibility = View.GONE
+                    } else {
+                        binding.emptyView.visibility = View.GONE
+                        binding.lista.visibility = View.VISIBLE
+                    }
+
+                })
+            }
+
+            R.id.Descendent -> {
+                database.Articles().getAllDesc().observe(this, {
+                    listArticles = it
+                    binding.lista.layoutManager =
+                        LinearLayoutManager(this@FragmentMain.requireContext())
+                    binding.lista.addItemDecoration(
+                        DividerItemDecoration(
+                            this@FragmentMain.requireContext(),
+                            DividerItemDecoration.VERTICAL
+                        )
+                    )
+
+                    val adapter = ArticleAdapter(listArticles)
+
+                    binding.lista.adapter = adapter
+                    if (listArticles.isEmpty()) {
+                        binding.emptyView.visibility = View.VISIBLE
+                        binding.lista.visibility = View.GONE
+                    } else {
+                        binding.emptyView.visibility = View.GONE
+                        binding.lista.visibility = View.VISIBLE
+                    }
+
+                })
+            }
+        }
+
+        return super.onOptionsItemSelected(item)
+    }
 
 
 }
